moved methods from writeFile to fileSystem
